Celulares y documentos: enteros grandes (integer out of range)
Un celular colombiano de 10 digitos (3209028064) supera el maximo de un
entero de 32 bits (2.147.483.647). En PostgreSQL el registro fallaba con
"integer out of range" y el usuario recien creado se borraba. Afectaba a
arrendatario, proveedor, coarrendatario y referencias (celular, celularDos);
administrador y propietario ya usaban enteros grandes para el celular.

celular, celularDos y doc_identificacion pasan a PositiveBigIntegerField (el
mismo tipo que ya usan dependientes y las cuentas bancarias). La migracion
solo ensancha la columna, no pierde datos.

+++ inmobiliarias/arrendatarios/models.py
@@ -32,10 +32,10 @@ class coarrendatario(models.Model):
     first_name = models.CharField(max_length=50, blank=True, null=True)  
     last_name = models.CharField(max_length=50, blank=True, null=True) 
     tipo_documento = models.CharField(max_length=3, choices=TipoDocumento.choices, blank=True, null=True)
-    doc_identificacion = models.IntegerField(blank=True, null=True)
+    doc_identificacion = models.PositiveBigIntegerField(blank=True, null=True)
     lugarExpCedula = models.CharField(blank=True,max_length=255)
     parentezco = models.CharField(max_length=30, blank=True, null=True) #Conyuge, Codeudor, Familiar o amigo
-    celular = models.IntegerField(blank=True, null=True)
+    celular = models.PositiveBigIntegerField(blank=True, null=True)
     email = models.EmailField(max_length=50, blank=True, null=True) 
     direccion = models.CharField(blank=True,max_length=255)
     barrio = models.CharField(blank=True,max_length=255)
@@ -50,7 +50,7 @@ class coarrendatario(models.Model):
     barrioCorrespondencia = models.CharField(blank=True,max_length=255)
     ciudadCorrespondencia = models.CharField(blank=True,max_length=255) 
     
-    celularDos = models.IntegerField(null=True, blank=True) 
+    celularDos = models.PositiveBigIntegerField(null=True, blank=True) 
 
     def __str__(self):
         return f"{self.first_name} {self.last_name}" 
@@ -82,7 +82,7 @@ class referencias(models.Model):
     first_name = models.CharField(max_length=30, blank=True, null=True)
     last_name = models.CharField(max_length=30, blank=True, null=True)    
     parentezco = models.CharField(max_length=30, blank=True, null=True)
-    celular = models.IntegerField(blank=True, null=True)    
+    celular = models.PositiveBigIntegerField(blank=True, null=True)    
     dir_residencia = models.CharField(max_length=50, blank=True, null=True)
     ocupacion = models.CharField(max_length=30, blank=True, null=True)
     motivoRetiro = models.CharField(max_length=50, blank=True, null=True)
